Use new Computest branding
Move the "Preferences" link in the navigation bar to the Profile page
because it messes up the navigation layout on small screens. The
navigation bar would then no longer fit on one line. Also, the useless
vertical bar in the navigation is removed, which would show up as a menu
item on small screens.

This also fixes header display on small screens. The header size will
now scale down on small screens.