Update Computest plugin for CTFd 2.1.1
Make the Computest plugin compatible with the new version of CTFd. The
theme files were copy-pasted from the core theme, so this removes all
theme customizations. Theme customizations will be restored in a
separate commit, so that it's easier to view and re-apply the theme
changes.

All functionality of the original plugin are restored, except for the
JSON API related changes. It appears that it's no longer possible to
overwrite existing APIs, as these are no longer implemented as standard
views. As a temporary solution, the original
`CTFd.utils.scores.get_standings()` function is modified.